Relocating Identicon component
