+++ src/ui/library/index.ts
@@ -44,4 +44,3 @@ export {Calendar} from './Calendar';
 export {Tabs, TabScreen, useTabIndex, useTabNavigation} from 'react-native-paper-tabs';
 export {useBottomSheet, BottomSheetTextInput, useBottomSheetInternal} from './BottomSheet';
 export {Skeleton} from './Skeleton';
-export {Identicon} from './Identicon';
