+++ src/ui/screens/AccountsScreen.tsx
@@ -1,19 +1,7 @@
 import React from 'react';
 import {FlatList, View, StyleSheet, TouchableOpacity, Keyboard} from 'react-native';
 import {NavigationProp} from '@react-navigation/native';
-import {
-  useTheme,
-  Divider,
-  IconButton,
-  List,
-  FAB,
-  Caption,
-  Menu,
-  Subheading,
-  Icon,
-  useBottomSheet,
-  Identicon,
-} from '@ui/library';
+import {useTheme, Divider, IconButton, List, FAB, Caption, Menu, Subheading, Icon, useBottomSheet} from '@ui/library';
 import SafeView, {noTopEdges} from '@ui/components/SafeView';
 import {CompleteNavigatorParamList} from '@ui/navigation/navigation';
 import {accountsScreen, importAccountScreen, mnemonicScreen, myAccountScreen} from '@ui/navigation/routeKeys';
@@ -27,6 +15,7 @@ import {AddExternalAccount} from '@ui/components/Account/AddExternalAccount';
 import type {AccountMeta, KeyringAccount} from 'polkadot-api';
 import {useAppAccounts} from '@polkadotApi/useAppAccounts';
 import {useKeyring} from '@polkadotApi/useKeyring';
+import {Identicon} from '@ui/components/Identicon';
 
 type Props = {
   navigation: NavigationProp<CompleteNavigatorParamList, typeof accountsScreen>;
