+++ src/ui/screens/Council/CouncilScreen.tsx
@@ -10,18 +10,8 @@ import {SelectAccount} from '@ui/components/SelectAccount';
 import {useCouncil, CouncilMember, Council} from 'src/api/hooks/useCouncil';
 import {useFormatBalance} from 'src/hooks/useFormatBalance';
 import {candidateScreen} from '@ui/navigation/routeKeys';
-import {
-  Button,
-  Divider,
-  Modal,
-  useTheme,
-  Caption,
-  Subheading,
-  Text,
-  TextInput,
-  HelperText,
-  Identicon,
-} from '@ui/library';
+import {Button, Divider, Modal, useTheme, Caption, Subheading, Text, TextInput, HelperText} from '@ui/library';
+import {Identicon} from '@ui/components/Identicon';
 import {Padder} from '@ui/components/Padder';
 import globalStyles, {standardPadding} from '@ui/styles';
 import {MotionsScreen} from './MotionsScreen';
