+++ src/ui/screens/ImportAccountScreen.tsx
@@ -6,7 +6,8 @@ import {ProgressBar} from '@ui/components/ProgressBar';
 import SafeView, {noTopEdges} from '@ui/components/SafeView';
 import {AccountsStackParamList} from '@ui/navigation/navigation';
 import {accountsScreen} from '@ui/navigation/routeKeys';
-import {Button, List, TextInput, useTheme, HelperText, Identicon} from '@ui/library';
+import {Button, List, TextInput, useTheme, HelperText} from '@ui/library';
+import {Identicon} from '@ui/components/Identicon';
 import {Padder} from '@ui/components/Padder';
 import globalStyles, {monofontFamily, standardPadding} from '@ui/styles';
 import zxcvbn from 'zxcvbn';
