+++ src/ui/screens/ImportAccountWithJsonFileScreen.tsx
@@ -7,7 +7,8 @@ import DocumentPicker, {DocumentPickerResponse} from 'react-native-document-pick
 import RNFS from 'react-native-fs';
 import {AccountsStackParamList} from '@ui/navigation/navigation';
 import {accountsScreen} from '@ui/navigation/routeKeys';
-import {Button, Caption, List, Text, TextInput, useTheme, Identicon} from '@ui/library';
+import {Button, Caption, List, Text, TextInput, useTheme} from '@ui/library';
+import {Identicon} from '@ui/components/Identicon';
 import {ErrorText} from '@ui/components/ErrorText';
 import {Padder} from '@ui/components/Padder';
 import globalStyles, {monofontFamily, standardPadding} from '@ui/styles';
