+++ src/ui/screens/MyIdentityScreen/ManageIdentityScreen.tsx
@@ -2,7 +2,8 @@ import React, {useCallback} from 'react';
 import {Alert, StyleSheet, View} from 'react-native';
 import {NavigationProp, RouteProp} from '@react-navigation/native';
 import {Layout} from '@ui/components/Layout';
-import {Button, List, Icon, Caption, Divider, IconButton, useBottomSheet, Identicon} from '@ui/library';
+import {Button, List, Icon, Caption, Divider, IconButton, useBottomSheet} from '@ui/library';
+import {Identicon} from '@ui/components/Identicon';
 import {useNetwork} from '@atoms/network';
 import IdentityInfoForm from '@ui/components/IdentityInfoForm';
 import InfoBanner from '@ui/components/InfoBanner';
