+++ src/ui/screens/Parachains/ParathreadsScreen.tsx
@@ -3,7 +3,8 @@ import {View, FlatList, StyleSheet, Linking} from 'react-native';
 import globalStyles, {standardPadding} from '@ui/styles';
 import SafeView, {noTopEdges} from '@ui/components/SafeView';
 import {useParathreads} from 'src/api/hooks/useParaThreads';
-import {List, Text, Divider, Subheading, Identicon} from '@ui/library';
+import {List, Text, Divider, Subheading} from '@ui/library';
+import {Identicon} from '@ui/components/Identicon';
 import {EmptyView} from '@ui/components/EmptyView';
 import LoadingView from '@ui/components/LoadingView';
 import type {Parathread} from 'src/api/hooks/useParaThreads';
