+++ src/ui/screens/SubIdentities/RegisterSubIdentitiesScreen.tsx
@@ -2,7 +2,8 @@ import React, {useEffect, useState} from 'react';
 import {Alert, FlatList, View} from 'react-native';
 import {stringShorten} from '@polkadot/util';
 import {NavigationProp, RouteProp} from '@react-navigation/native';
-import {Button, Caption, Subheading, List, Divider, IconButton, useTheme, useBottomSheet, Identicon} from '@ui/library';
+import {Button, Caption, Subheading, List, Divider, IconButton, useTheme, useBottomSheet} from '@ui/library';
+import {Identicon} from '@ui/components/Identicon';
 import {Layout} from '@ui/components/Layout';
 import SafeView, {noTopEdges} from '@ui/components/SafeView';
 import {AccountsStackParamList} from '@ui/navigation/navigation';
